Adds a maximum fee variable
In order to prevent a large fee being paid this commit adds a max fee property which will cause the send command to abort if the built route exceeds it.

The purpose is to prevent unexpectedly large fees being paid which can occur if relay fees are changed on a peer in between checking the build command and then running the send command.